refactor(api): extract service layer logic from route handlers
