+++ src/routes/api/status/status.validation.ts
@@ -1,0 +1,5 @@
+import { z } from "zod";
+
+export const getStatusValidation = z.object({});
+
+export type GetStatusType = z.infer<typeof getStatusValidation>;
